Code refactoring and comments added

+++ inc/libraries_mockup.h
@@ -11,11 +11,11 @@
 //~ Function : gyrometerAcq
 //~ ----------------------------
 //~ Mockup of the gyrometer acquisition function
-//~ inout : float &yawRate, is the returned yaw rate in rads/s, uint64_t &timestamp
+//~ inout : float &yawRate, is the returned yaw rate in rads/s, uint32_t &timestamp
 //~  is the time in miliseconds at which the yaw rate acquisition was taken
 //~
 //~ output : int is 1 if suceess, -1 if error
-int gyrometerAcq(float &yawRate, uint64_t &timestamp);
+int gyrometerAcq(float &yawRate, uint32_t &timestamp);
 
 //~ Function : odometryAcq
 //~ ----------------------------
@@ -23,8 +23,8 @@ int gyrometerAcq(float &yawRate, uint64_t &timestamp);
 //~ inout : std::array<float, 4> &odometry, is the returned wheel odometry in
 //~   meters with the following order :
 //~           [left_back, right_back, left_front, right_front]
-//~   uint64_t &timestamp is the time in miliseconds at which the wheel odometry
+//~   uint32_t &timestamp is the time in miliseconds at which the wheel odometry
 //~     was taken
 //~
 //~ output : int is 1 if suceess, -1 if error
-int odometryAcq(std::array<float, 4> &odometry, uint64_t &timestamp);
+int odometryAcq(std::array<float, 4> &odometry, uint32_t &timestamp);
