+++ src/libraries_mockup.cpp
@@ -1,10 +1,10 @@
 
 #include "libraries_mockup.h"
 
-int gyrometerAcq(float &yawRate, uint64_t &timestamp){
+int gyrometerAcq(float &yawRate, uint32_t &timestamp){
   return 1;
 }
 
-int odometryAcq(std::array<float, 4> &odometry, uint64_t &timestamp){
+int odometryAcq(std::array<float, 4> &odometry, uint32_t &timestamp){
   return 1;
 }
